refactor(runner): consolidate scheduler facade (scheduler-v2 3f)
