+++ packages/runner/test/action-fingerprint.test.ts
@@ -1,7 +1,7 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 
-import { schedulerImplementationFingerprint } from "../src/scheduler/action-run.ts";
+import { schedulerImplementationFingerprint } from "../src/scheduler/run.ts";
 import type { Action } from "../src/scheduler/types.ts";
 
 function makeAction(props: Record<string, unknown>): Action {
